+++ lib/latlong/calculator/haversine.dart
@@ -17,7 +17,13 @@
  * limitations under the License.
  */
 
-part of latlong;
+import 'dart:math' as math;
+
+import 'package:validate/validate.dart';
+
+import '../../latlong.dart';
+import '../interfaces.dart';
+import '../lat_lng.dart';
 
 class Haversine implements DistanceCalculator {
   const Haversine();