+++ lib/spline.dart
@@ -25,9 +25,4 @@
 ///
 /// Java way: [A Catmull Rom Spline (curve) Implementation in Java](http://hawkesy.blogspot.co.at/2010/05/catmull-rom-spline-curve-implementation.html)
 ///
-library spline;
-
-import 'package:validate/validate.dart';
-//import 'package:logging/logging.dart';
-
-part 'spline/CatmullRomSpline.dart';
+export 'spline/catmull_rom_spline.dart';